canapy-old/__init__.py
import reservoirpy as rpy

from . import sequence

from .dataset import Dataset, Config
from .model import Model, NSynModel, SynModel, Annotator, Trainer, Ensemble
from .processor import Processor

from .config import default_config

rpy.verbosity(0)

__all__ = [
    "Dataset",
    "Config",
    "Model",
    "NSynModel",
    "SynModel",
    "Ensemble",
    "Trainer",
    "Annotator",
    "Processor",
    "sequence",
    "default_config",
]
