pyproject.toml
[tool.poetry]
name = "canapy-reborn"
version = "0.1.0"
description = ""
authors = ["nTrouvain <nathan.trouvain@inria.fr>"]
readme = "README.md"
packages = [{include = "canapy_reborn"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numba = "^0.57.1"
librosa = "^0.10.0.post2"
scikit-learn = "^1.2.2"
scipy = "^1.11.0"
numpy = "<=1.25"
matplotlib = "^3.7.1"
joblib = "^1.2.0"
pandas = "^2.0.2"
crowsetta = "^5.0.1"
attrs = "^23.1.0"
pandera = "^0.15.1"
pyyaml = "^6.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
